feat: finalizando visualização dos dados
